+++ Reflection/test/utopia/reflection/test/AnimatedTransitionTest.scala
@@ -6,12 +6,9 @@ import utopia.genesis.color.Color
 import utopia.genesis.generic.GenesisDataType
 import utopia.genesis.handling.mutable.ActorHandler
 import utopia.genesis.shape.Axis.Y
-import utopia.genesis.shape.Direction1D.Negative
 import utopia.reflection.component.drawing.immutable.BorderDrawer
-import utopia.reflection.component.swing.AnimatedTransition
+import utopia.reflection.component.swing.AnimatedVisibility
 import utopia.reflection.component.swing.label.TextLabel
-import utopia.reflection.container.swing.Stack.AwtStackable
-import utopia.reflection.container.swing.SwitchPanel
 import utopia.reflection.container.swing.window.Frame
 import utopia.reflection.container.swing.window.WindowResizePolicy.Program
 import utopia.reflection.localization.{Localizer, NoLocalization}
@@ -47,34 +44,28 @@ object AnimatedTransitionTest extends App
 	implicit val baseContext: ComponentContext = baseCB.result
 	
 	// Creates the component to display
-	val originComponent = TextLabel.contextual("This is to be drawn as image")
+	val originComponent = TextLabel.contextual("I'm Animated :)")
 	originComponent.addCustomDrawer(new BorderDrawer(Border.square(2, Color.blue)))
 	originComponent.addResizeListener { e => println(s"Component size changed: $e") }
 	
 	// Creates the transition components
-	val appearance = new AnimatedTransition(originComponent, Y)
-	actorHandler += appearance
-	
-	val disappearance = new AnimatedTransition(originComponent, Y, Negative)
-	actorHandler += disappearance
-	
-	// Wraps the component and transitions to a managed switch panel
-	val mainPanel = SwitchPanel[AwtStackable](appearance)
-	val content = mainPanel.framed(16.downscaling.square, Color.white)
+	val transitionWrapper = new AnimatedVisibility(originComponent, actorHandler, Y)
+	val content = transitionWrapper.framed(16.downscaling.square, Color.white)
 	
 	// Starts test
 	val frame = Frame.windowed(content, "Animated Transition Test", Program)
 	frame.setToCloseOnEsc()
 	new SingleFrameSetup(actorHandler, frame).start()
 	
+	// Displays the component after 0.5 seconds
 	WaitUtils.delayed(0.5.seconds) {
-		appearance.start().foreach { _ =>
-			println("Component fully appeared")
-			mainPanel.set(originComponent)
-			WaitUtils.delayed(3.seconds) {
-				println("Component starting to disappear")
-				mainPanel.set(disappearance)
-				disappearance.start().foreach { _ => frame.close() }
+		transitionWrapper.isShown = true
+		// Hides it again after 5 seconds
+		WaitUtils.delayed(5.seconds) {
+			transitionWrapper.isShown = false
+			// Finally closes program after 1 second of disappearance
+			WaitUtils.delayed(1.seconds) {
+				frame.close()
 			}
 		}
 	}